+++ docs/documentation/type._editables.*.image_options.test.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.test
+url: /configuration-file/types/_editables/*/image_options/test/
+title: ''
+description: ''
+examples: []
+show_in_navigation: false
